Move Marker to separate module
